requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_classic
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
langchain-tavily